OOP start
This is the start. It's not finished and it doesn't work yet (player isn't showing on screen). You guys can start adding pips and walls, just ignore the player class for now and do it in a new branch. If any of you feel like working on it go for it. I'll keep going later but I need a break.